# dist/openapi.yaml
openapi: 3.0.0
info:
  title: Recipes api
  description: API for recipes project
  version: 1.0.1
servers:
# local:
- url: "http://127.0.0.1:3000"
# remote:
# - url: "http://132.72.65.211"
  description: SwaggerHub API Auto Mocking hey
...
paths:
...
  /recipes/{recipeId}: #example: /recipes/716429
    get:
      tags:
      - Recipes
      summary: Return the preview details on recipe by his ID
      description: Get preview details on recipe by hid ID
      parameters:
      - name: recipeId
        in: path
        description: recipe ID
        required: true
        style: simple
        explode: false
        schema:
          type: integer
      responses:
        "200":
          description: data about the recipe returned
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RecipePreview'
                  watched:
                    type: boolean
                    description: if a user is logged in, return whether the recipe was watched or not, if not logged in return false
                    example: "1"
                  favorite:
                    type: boolean
                    description: if a user is logged in, return whether the recipe was favorited or not, if not logged in return false
                    example: "1"
        "404":
          description: recipeId is not found
  /recipes/searchRecipes:
    post: #the access to the api was only allowed when it's a post request, get request returned an error
      tags:
      - Recipes
      summary: search recipes by filters
      description: search recipes by filters
      # parameters:
      # - name: query
      #   description: search query for API
      #   required: false
      #   style: simple
      #   explode: false
      #   example: "pasta"
      #   schema:
      #     type: string
          
      # - name: cuisine
      #   description: Select certain cuisine's recipes
      #   required: false
      #   style: simple
      #   explode: false
      #   example: "Italian"
      #   schema:
      #     type: string
          
      # - name: diet
      #   description: Select certain diet recipes
      #   required: false
      #   style: simple
      #   explode: false
      #   example: "vegetarian"  
      #   schema:
      #     type: string
                
      # - name: intolerances
      #   description: Select certain intolerances recipes
      #   required: false
      #   style: simple
      #   explode: false
      #   example: "Peanut"  
      #   schema:
      #     type: string
               
      # - name: number
      #   enum: [5,10,15]
      #   description: Select amount of recipes to be returend
      #   required: true
      #   default: 5
      #   style: simple
      #   explode: false
      #   example: 10   
      #   schema:
      #     type: integer
             
      # - name: watched
      #   type: boolean
      #   description: if a user is logged in, return whether the recipe was watched or not, if not logged in return false
      #   example: "1"
      # - name: favorite
      #   type: boolean
      #   description: if a user is logged in, return whether the recipe was favorited or not, if not logged in return false
      #   example: "1"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                query:
                  type: string
                  description: search query for API
                  example: pasta
                cuisine:
                  type: string
                  description: select certain cuisine's recipes
                  example: "Italian"
                diet:
                  type: string
                  description: select certain diet recipes
                  example: "vegetarian"
                intolerances:
                  type: string
                  description: select certain intolerances recipes
                  example: "Peanut"
                number:
                  type: integer
                  enum: [5,10,15]
                  default: 5
                  description: select amount of recipes to be returned
                  example: "10"
                watched:
                  type: boolean
                  description: if a user is logged in, return whether the recipe was watched or not, if not logged in return false
                  example: "0"
                favorite:
                  type: boolean
                  description: if a user is logged in, return whether the recipe was favored or not, if not logged in return false
                  example: "0"
      responses:
        "200":
          description: The Recipe successfully saved as watched
        "422":
          description: Invalid or missing arguments
  /recipes/randomRecipes:
    post: #the access to the api was only allowed when it's a post request, get request returned an error
      tags:
      - Recipes
      summary: get random recipes
      description: get random recipes
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                $ref: '#/components/schemas/RecipePreview'
                watched:
                  type: boolean
                  description: if a user is logged in, return whether the recipe was watched or not, if not logged in return false
                  example: "1"
                favorite:
                  type: boolean
                  description: if a user is logged in, return whether the recipe was favorited or not, if not logged in return false
                  example: "1"
      responses:
        "200":
          description: The Recipe successfully saved as watched
...
components:
  schemas:
...
    RecipePreview:
      type: object
      properties:
        id:
          type: integer
          description: recipe identifier
          example: 102
        title:
          type: string
          description: recipe name
          example: Simple Macaroni and Cheese
        image:
          type: string
          description: recipe image
          format: URL
          example: https://spoonacular.com/recipeimages/655455-566x370.jpg
        readyInMinutes:
          type: integer
          description: the time takes to make this recipe
          example: 45
        popularity:
          type: integer
          description: how many likes got this recipe
          example: 5987
        vegetarian:
          type: boolean
          description: indication if the recipe is vegetarian
          example: false
        vegan:
          type: boolean
          description: indication if the recipe is vegan
          example: false
        glutenFree:
          type: boolean
          description: indication if the recipe is gluten free
          example: false
...